requirements.txt
python-telegram-bot
requests
python-dotenv
mysql-connector-python 